Refactor React app configuration.
* Backup original notes and examples in config/example.js.
* Move config code out of src/app.js and into various config/ files
(this is a more clear organizational pattern in my opinion).
* Move configuration of Apollo middleware outside of base config class;
config class should probably be kept blank with calls made to its api
to make changes.
